requirements.txt
pandas~=2.3.3
numpy~=1.26.4
scikit-learn~=1.8.0
scikit-surprise~=1.1.4
matplotlib~=3.10.8